refactor : 멘토 지원 위치 조정
