+++ src/main/java/com/dementor/domain/admin/controller/AdminMentorApplymentController.java
@@ -1,11 +1,11 @@
 package com.dementor.domain.admin.controller;
 
-import com.dementor.domain.admin.dto.request.ApplymentRejectRequest;
-import com.dementor.domain.admin.dto.response.ApplymentApprovalResponse;
-import com.dementor.domain.admin.dto.response.ApplymentDetailResponse;
-import com.dementor.domain.admin.dto.response.ApplymentRejectResponse;
-import com.dementor.domain.admin.dto.response.ApplymentResponse;
-import com.dementor.domain.admin.service.AdminMentorApplymentService;
+import com.dementor.domain.mentorapplyproposal.dto.request.ApplymentRejectRequest;
+import com.dementor.domain.mentorapplyproposal.dto.response.ApplymentApprovalResponse;
+import com.dementor.domain.mentorapplyproposal.dto.response.ApplymentDetailResponse;
+import com.dementor.domain.mentorapplyproposal.dto.response.ApplymentRejectResponse;
+import com.dementor.domain.mentorapplyproposal.dto.response.ApplymentResponse;
+import com.dementor.domain.mentorapplyproposal.service.AdminMentorApplymentService;
 import com.dementor.global.ApiResponse;
 import com.dementor.global.common.pagination.PaginationUtil;
 import com.dementor.global.common.swaggerDocs.AdminMentorApplymentSwagger;
@@ -16,12 +16,13 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/admin/mentor/applyment")
 @RequiredArgsConstructor
-// TODO : 관리자만 할 수 있는 설정
+@PreAuthorize("hasRole('ADMIN')")
 public class AdminMentorApplymentController implements AdminMentorApplymentSwagger {
 
     private final AdminMentorApplymentService adminMentorApplymentService;
