+++ src/main/java/com/dementor/domain/mentorapplyproposal/dto/response/ApplymentRejectResponse.java
@@ -1,4 +1,4 @@
-package com.dementor.domain.admin.dto.response;
+package com.dementor.domain.mentorapplyproposal.dto.response;
 
 import java.time.LocalDateTime;
