+++ src/main/java/com/dementor/domain/mentorapplyproposal/dto/response/ApplymentResponse.java
@@ -1,4 +1,4 @@
-package com.dementor.domain.admin.dto.response;
+package com.dementor.domain.mentorapplyproposal.dto.response;
 
 import com.dementor.domain.job.entity.Job;
 import com.dementor.domain.mentorapplyproposal.entity.MentorApplyProposal;
