+++ src/main/java/com/dementor/domain/mentorapplyproposal/dto/response/AttachmentInfo.java
@@ -1,4 +1,4 @@
-package com.dementor.domain.admin.dto.response;
+package com.dementor.domain.mentorapplyproposal.dto.response;
 
 import com.dementor.domain.postattachment.entity.PostAttachment;
